Bring README.md up-to-date
This updates the README to provide a link to running it with GOV.UK
docker and removes the VM specific commands. As startup.sh is removed it
also removes the suggestion on how to run this locally with a custom
docker set-up.

This also adds a licence that was missing and makes the title more
consistent.

There were previously two commands listed before that should be used before
running the app: `bundle exec rake data_import:all` and `bundle exec rake
search:index`. I found that I could run the app ok if I only ran the
search task (without it the app crashed on the 2nd page). I tried
running the `data_import:all` task numerous times but it would
consistently grind to a halt after about 20 minutes of running and
reach a glacial pace of import. I couldn't work out quite why this is,
but it seems to be related to mongo many-to-many relations. I felt the
best thing to do is to not include a suggestion to run it since it
doesn't seem 100% necessary to run the app and I didn't want to give
the next dev who tries to run this a miserable experience.